+++ MdpForgot.php
@@ -98,18 +98,18 @@ if ($result > 0)  {                                                             
 
 <nav class="navbar navbar-expand-lg bg-dark mb-5">
             <div class="container-fluid">
-                <a class="navbar-brand text-white" href="#">MyBiblio</a>
+                <a class="navbar-brand text-white" >MyBiblio</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
                 <div class="collapse navbar-collapse " id="navbarSupportedContent">
                     <ul class="navbar-nav me-auto mb-2 mb-lg-0">
                         <li class="nav-item">
-                            <a class="nav-link active text-white" aria-current="page" href="#"></a>
+                            <a class="nav-link active text-white" aria-current="page" ></a>
                         </li>
 
                         <li class="nav-item">
-                            <a class="nav-link text-white" href="index.php"></a>
+                            <a class="nav-link text-white" ></a>
                         </li>
                     </ul>
