+++ dao.php
@@ -66,8 +66,9 @@ class DAO
 	}
 
 
-    //CLEMENT
+    //--------------------------------------------------------------------------------------------CLEMENT -------------------------------------------------------------------------------------
     
+
     //FONCTION QUI RECUPERE LES RESULTATS DES REQUETES SQL
     public function getResultat($requete)
     {
@@ -84,7 +85,7 @@ class DAO
             return $declaration->fetchAll();
         }
     }
-
+    //FONCTION POUR RECUPERER UN SEUL ELEMENT DE LA REQUETE
     public function getAlone($requete, $param)
     {
 
@@ -110,7 +111,7 @@ class DAO
     }
 
 
-
+    //FONCTION POUR RECUPERER LES LIVRES
     function get_livre(){
 
         $sql="SELECT id_livre, titre_livre, disponibilite_id FROM `livres`";
@@ -118,7 +119,7 @@ class DAO
         
     }
 
-
+    //FONCTION POUR RECUPERER le titre du livre avec GETALONE pour la fonction emprunt
     function get_livre_emprunt($param = []){
 
         $sql="SELECT * FROM `livres` where titre_livre = :inputTitre";
@@ -126,6 +127,13 @@ class DAO
         
     }
 
+     //FONCTION POUR RECUPERER le titre du livre avec GETALONE pour la fonction rendu
+    function get_livre_rendu($param = []){
+
+        $sql="SELECT * FROM `livres` where titre_livre = :inputTitre";
+        return $this->getAlone($sql, $param);
+        
+    }
 
     //FONCTION POUR RECUPERER LES GENRES DE LIVRES
     function getGenre(){
@@ -143,14 +151,24 @@ class DAO
 
     }
 
+    //FONCTION POUR RECUPERER LE NOMBRE DE LIVRES EN STOCK
+    function getStock($param= []){
+        
+    $sql="SELECT Nombre_livre FROM stock WHERE id_livre = :id_livre ";
+    return $this->getAlone($sql,$param);
+
+    }
 
 
-    function get_livre_utilisateur($name){
+    //FONCTION POUR RECUPERER L'ID DU LIVRE DANS LA TABLE LIVRE_UTILISATEUR
+    function get_livre_utilisateur(){
 
         $sql="SELECT id_livre FROM `livre_utilisateur`";
         return $this->getResultat($sql);
 
     }
+
+    //FONCTION POUR RECUPERER LA DISPO DU LIVRE QUI CORRESPOND AU LIVRE SOUHAITE
     function verif_dispo_livre($id_livre){
 
         $sql="SELECT disponibilite_id FROM livres WHERE id_livre LIKE  $id_livre";
@@ -158,12 +176,18 @@ class DAO
 
     }
 
-    function getUtilisateur(){
 
-        $sql= "SELECT * FROM `utilisateurs`";
+    function verif_dispo_livre2($id_livre = []){
+
+        $sql="SELECT disponibilite_id FROM livres WHERE id_livre LIKE  $id_livre";
         return $this->getResultat($sql);
 
     }
+
+    //FONCTION POUR TOUT RECUPERER DE LA TABLE UTILISATEUR
+    // function getUtilisateur(){
+
+
 
 
     //FONCTION POUR RECUPERER LES NOMS DES AUTEURS 
@@ -287,24 +311,43 @@ class DAO
     }
 
 
-
-
     function emprunt_livre($param){
 
 
         if(isset($_POST['liste_livre_emprunt'])){
 
-           
-        
+            //AJout de la date d'emprunt et de la date de retour 15 jours apres 
+            $dateEmprunt= date("Y-m-d");
+            $dateRetour = date('Y-m-d', strtotime($dateEmprunt. ' + 15 days'));
             $idlivre = $param;
+            $id_util = $_POST['utilisateur'];
+
+        
+
+            //Récupération du nombre de livre, dans la table STOCK, par ID de livre
+            //SI le nombre de livres est différent de 0 
+            if($this->getStock(["id_livre" =>$param])['Nombre_livre'] != 0 ){
+
+            //Mise a jour de la table stock a l'endroit qui correspond a l'ID du livre
+            $sql2="UPDATE stock SET Nombre_livre = Nombre_livre-1 WHERE id_livre = $idlivre";
+            $this->bdd->query($sql2);
+
+            //INSERTION dans la table livre utilisateur
+            $sql1="INSERT INTO livre_utilisateur (`id_livre`,`id_utilisateur`, `date_emprunt`,`date_retour`) VALUES (?,?,?,?)";
+            $query = $this->bdd->prepare($sql1);
+            $query->execute([$idlivre, $id_util,$dateEmprunt,$dateRetour]);
+
+            }
+            
+            //SI le nombre de livre est 0
+            if($this->getStock(["id_livre" =>$param])['Nombre_livre'] == 0 ){
+            //Mise a jour de la table livres, on change la disponibilité du livre en fonction de nombre de livres en stock
             $sql="UPDATE livres SET disponibilite_id = 1 WHERE id_livre = $idlivre";
             $this->bdd->query($sql);
-            $id_util = $_POST['utilisateur'];
-            $sql1="INSERT INTO livre_utilisateur (`id_livre`,`id_utilisateur`) VALUES (?,?)";
-            $query = $this->bdd->prepare($sql1);
-            $query->execute([$idlivre, $id_util]);
+                
 
-            
+                }
+
             }
                
         }
@@ -315,17 +358,23 @@ class DAO
             if(isset($_POST['liste_livre_rendu'])){
    
                 $idlivre = $param;
-                $sql="UPDATE livres SET disponibilite_id = 0 WHERE id_livre = $idlivre";
-                $this->bdd->query($sql);
                 $id_util = $_POST['utilisateur'];
 
+                //Supression de la ligne qui correspond a l'emprunt du livre dans la table livre_utilisateur
                 $sql1="DELETE FROM livre_utilisateur WHERE id_livre =  $idlivre AND id_utilisateur = $id_util";
                 $this->bdd->query($sql1);
+                //Mise a jour du stock lrs d'un emprunt de livre
+                $sql2="UPDATE stock SET Nombre_livre = Nombre_livre+1 WHERE id_livre = $idlivre";
+                $this->bdd->query($sql2);
+                //SI le stock est supérieur à 1 on modifie la disponibilité du livre
+                if($this->getStock(["id_livre" =>$param])['Nombre_livre']  >= 1 ){
+                $sql="UPDATE livres SET disponibilite_id = 0 WHERE id_livre = $idlivre";
+                $this->bdd->query($sql);
 
                 }
                    
             }
-
+        }
         
     function suppr_livre($idlivre){
 
@@ -335,24 +384,60 @@ class DAO
 
     }
 
+
+        
+//-----------------------------------------------------------------------------------------------JASON-------------------------------------------------------------------------------------------
+	
+
 //JASON
-	/* méthode qui renvoit tous les résultats sous forme de tableau
-	*/
+	/* méthode qui renvoit tous les résultats sous forme de tableau*/
+
 	public function getLivre() {
-		$sql="SELECT image, titre_livre, isbn, shortDescription,id_livre FROM livres;";
+		$sql="SELECT livres.image, livres.titre_livre, livres.isbn, genres.nom_genre, livres.id_livre, livres.shortDescription, auteurs.nom_auteur
+        FROM livres
+        INNER JOIN livres_genres ON livres.id_livre = livres_genres.livre_id
+        INNER JOIN genres ON livres_genres.genre_id = genres.id_genre
+        
+        INNER JOIN livre_auteur ON livres.id_livre = livre_auteur.id_livre
+        INNER JOIN auteurs ON livres.auteur_id = auteurs.id_auteur;";
 		return $this->getResults($sql);
 	}
-	
+	//Methode pour le bouton de suppression sur la page catalogue de livre
 	public function getDelete() {
 		$sql="SELECT image, titre_livre, isbn, shortDescription,id_livre FROM livres;";
 		return $this->getResults($sql);
 	}
-	
+	//info utilisateur sur modal page_utilisateur
+    public function getUtilisateurLivreEmprunte() {
+        $sql="SELECT nom_utilisateur, prenom_utilisateur,identifiant_utilisateur, utilisateurs.id_utilisateur,titre_livre,date_emprunt,date_retour 
+        FROM `utilisateurs` 
+        INNER JOIN livre_utilisateur ON utilisateurs.id_utilisateur= livre_utilisateur.id_utilisateur 
+        INNER JOIN livres ON livres.id_livre = livre_utilisateur.id_livre  
+        ORDER BY nom_utilisateur;";
+        return $this->getResults($sql);
+    }
+    // info utilisateur sur modal Sur index
+    public function getLivreEmprunteParUser() {
+        $sql="SELECT nom_utilisateur, prenom_utilisateur,identifiant_utilisateur, utilisateurs.id_utilisateur,titre_livre,livres.id_livre 
+        FROM `utilisateurs` 
+        INNER JOIN livre_utilisateur ON utilisateurs.id_utilisateur= livre_utilisateur.id_utilisateur 
+        INNER JOIN livres ON livres.id_livre = livre_utilisateur.id_livre 
+        ORDER BY nom_utilisateur;";
+        return $this->getResults($sql);
+    }
+    function getUtilisateur(){
+
+        $sql= "SELECT nom_utilisateur, prenom_utilisateur,identifiant_utilisateur,type_utilisateur,id_utilisateur 
+        FROM utilisateurs 
+        WHERE type_utilisateur !=1";
+        return $this->getResultat($sql);
+
+    }
 
 
+    //-----------------------------------------------------------------------------------------------------DAVID-------------------------------------------------------------------------------------
 
 
-//DAVID
     //fonction pour ajouter des utilisateurs depuis le formulaire d'inscription:
         
     //mettre en paramètre les données stockées en POST    
@@ -423,7 +508,9 @@ class DAO
 
 
 
- //PAUL 
+ //----------------------------------------------------------------------------------------PAUL-------------------------------------------------------------------------------------------------
+
+
    // fonction changment de status de la dispo dans datatable 
    
    public function statusDispo()
@@ -447,6 +534,51 @@ class DAO
     return $statusArray;
 }
 
+    //fonction ajout d'utilisateur 
+    public function ajoutUtilisateur($nom_utilisateur, $prenom_utilisateur, $mail_utilisateur, $tel_utilisateur)
+    {
+        $identifiant_utilisateur = 0;
+        $message = ''; // Variable pour stocker le message
+    
+        // Vérifie si l'identifiant utilisateur existe déjà
+        $sql_check_identifiant = "SELECT COUNT(*) FROM utilisateurs WHERE identifiant_utilisateur = ?";
+        $query_check_identifiant = $this->bdd->prepare($sql_check_identifiant);
+    
+        do {
+            // Génère un nouvel identifiant utilisateur aléatoire entre 10000 et 999999999
+            $identifiant_utilisateur = mt_rand(10000, 99999999);
+    
+            // Vérifie si le nouvel identifiant existe déjà 
+            $query_check_identifiant->execute([$identifiant_utilisateur]);
+            $identifiant_exists = $query_check_identifiant->fetchColumn();
+    
+            // Si l'identifiant a été réaffecté, stocke le message
+            if ($identifiant_utilisateur != mt_rand(10000, 999999999)) {
+                $message = "Identifiant réaffecté";
+            }
+        } while ($identifiant_exists);
+    
+        // Vérifie si l'e-mail existe déjà dans la base de données
+        $sql_check_email = "SELECT COUNT(*) FROM utilisateurs WHERE mail_utilisateur = ?";
+        $query_check_email = $this->bdd->prepare($sql_check_email);
+        $query_check_email->execute([$mail_utilisateur]);
+        $email_exists = $query_check_email->fetchColumn();
+    
+    if ($email_exists) {
+            // L'e-mail existe déjà, retourne un message d'erreur
+            return "L'utilisateur avec cet e-mail existe déjà.";
+        } else {
+            // L'e-mail n'existe pas, procède à l'insertion
+            $sql_insert_user = "INSERT INTO utilisateurs (nom_utilisateur, prenom_utilisateur, mail_utilisateur, tel_utilisateur, identifiant_utilisateur) VALUES (?, ?, ?, ?, ?)";
+            $query_insert_user = $this->bdd->prepare($sql_insert_user);
+            $query_insert_user->execute([$nom_utilisateur, $prenom_utilisateur, $mail_utilisateur, $tel_utilisateur, $identifiant_utilisateur]);
+    
+            // Retourne le message (peut être vide s'il n'y a pas eu de réaffectation d'identifiant)
+            return $message . " Utilisateur ajouté avec succès!";
+        }
+    }
+
 }
+
 ?>
