+++ suppr.php
@@ -1,5 +1,3 @@
-
-
 <?php 
 session_start();                                                       //on démarre la session pour pouvoir utiliser les variables de session
 require_once("dao.php");                                          //on fait la jonction avec le fichier DAO
@@ -13,7 +11,7 @@ if (isset($_POST['btn_suppr'])){
 
 
     $dao->suppr_livre($_POST['btn_suppr']); }
-header('location:index.php');    
+header('location:page_livre.php');    
 
 
 $dao->disconnect();
